refactored the code, imported FontAwesomeIcon instead using SVG files though the app
