--- src/components/Layout/Layout.styles.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const PageLayout = styled.div`
-  background-color: ${(props) => props.theme.background};
-  /* height: 100vh; */
-`;
