+++ src/contexts/CurrencyColorProvider/CurrencyColorProvider.jsx
@@ -5,10 +5,6 @@ export const CurrencyColorContext = createContext();
 export const CurrencyColorProvider = ({children}) => {
     const [currency, setCurrency] = useState("EUR")
     const [colorMode, setColorMode] = useState("dark")
-    // const [chartData, setChartData] = useState({})
-    // const [optionsData, setOptionsData] = useState({})
-
-
 
     return(
         <CurrencyColorContext.Provider value={{currency, setCurrency,colorMode,setColorMode,}}>
