--- src/utils/tableColors/tableColors.js
@@ -1,23 +1,0 @@
-export const TableLeftcolors = [
-  "#FFB528",
-  "#474C77",
-  "#1BA27A",
-  "#BB9F33",
-  "#FE7D43",
-  "#B3404A",
-  "#2775C9",
-  "#83808B",
-  "#345D9D",
-];
-
-export const TableRightcolors = [
-  "#FEE158",
-  "#8A92B2",
-  "#FFFFFF",
-  "#E4CD82",
-  "#FFDCCE",
-  "#F4B2B0",
-  "#FFFFFF",
-  "#F09242",
-  "#F09242",
-];
